chore(): decouple sw functions
